requirements.txt
asgiref==3.7.2
cloudinary==1.33.0
dj-database-url==2.0.0
dj-rest-auth==4.0.1
Django==3.2.20
django-allauth==0.54.0
django-cloudinary-storage==0.3.0
django-cors-headers==4.2.0
django-filter==23.2
django-rest-auth==0.9.5
djangorestframework==3.14.0
oauthlib==3.2.2
Pillow==10.0.0
PyJWT==2.8.0
python3-openid==3.2.0
pytz==2023.3
requests-oauthlib==1.3.1
sqlparse==0.4.4
urllib3==1.26.16
